backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
reportlab==4.0.8
PyMuPDF==1.24.14
python-jose[cryptography]==3.3.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv==1.0.0
PyJWT==2.8.0
types-PyJWT==1.7.1
stripe==7.11.0 